+++ CHANGELOG.md
@@ -35,6 +35,7 @@
   * Support setting the `internal` flag for created issue and MR notes.
   * Support the `WaitingForResource` job scope when filtering jobs.
   * Support the `include_retried` field for project jobs.
+  * Support filtering MRs by approver usernames.
 
 ## Changes